requirements.txt
Django==5.2.12
django-ratelimit==4.1.0
langchain==0.3.15
langchain-community==0.3.15
langchain-google-genai==2.0.8
langchain-text-splitters==0.3.6
faiss-cpu==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
google-generativeai==0.8.4
langchain-groq